Refactor types and config files to be outside of pages directory
