+++ pages/_app.tsx
@@ -1,11 +1,11 @@
 import "../styles/globals.css";
 import type { AppProps } from "next/app";
 import { Container, createStyles, MantineProvider } from "@mantine/core";
-import { CONFIG_WEBSITE_THEME } from "./config/global/App";
+import { CONFIG_WEBSITE_THEME } from "../config/global/App";
 import HeaderMenu from "./components/viewport/HeaderMenu";
-import { CONFIG_HEADERMENU } from "./config/viewport/HeaderMenu";
+import { CONFIG_HEADERMENU } from "../config/viewport/HeaderMenu";
 import FooterMenu from "./components/viewport/FooterMenu";
-import { CONFIG_FOOTERMENU } from "./config/viewport/FooterMenu";
+import { CONFIG_FOOTERMENU } from "../config/viewport/FooterMenu";
 
 const useStyles = createStyles(() => ({
   root: {
