+++ pages/components/viewport/FooterMenu.tsx
@@ -1,5 +1,5 @@
 import { Anchor, Container, createStyles, Flex, Text } from "@mantine/core";
-import { CONFIG_HEADER_FOOTER_HEIGHT } from "../../config/viewport/HeaderMenu";
+import { CONFIG_HEADER_FOOTER_HEIGHT } from "../../../config/viewport/HeaderMenu";
 import ColorStrip from "../utilities/ColorStrip";
 
 const useStyles = createStyles((theme) => ({
